requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
arxiv==2.1.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autogenstudio==0.1.2
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.127
botocore==1.34.127
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.2.1
fairscale==0.4.13
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
feedparser==6.0.10
filelock==3.13.3
fire==0.6.0
FLAML==2.1.2
Flask==3.0.3
flatbuffers==24.3.25
free_proxy==1.1.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
google==3.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.56.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
graphviz==0.20.3
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.4.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.2.1
langchain==0.1.15
langchain-anthropic==0.1.11
langchain-aws==0.1.3
langchain-community==0.0.38
langchain-core==0.2.9
langchain-google-genai==1.0.3
langchain-groq==0.1.3
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langgraph==0.1.1
langsmith==0.1.81
libclang==18.1.1
-e git+https://github.com/meta-llama/llama.git@54c22c0d63a3f3c9e77f43a6a3041c00018f4964#egg=llama
loguru==0.7.2
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
minify_html==0.15.0
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openai==1.34.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==23.2
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.0
playwright==1.43.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.19
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.29
pydantic==2.7.4
pydantic_core==2.18.4
pyee==11.1.0
Pygments==2.17.2
pymongo==4.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-version==0.0.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requirements.py==1.0.1
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
scrapegraphai==1.7.0
semchunk==1.0.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlmodel==0.0.19
starlette==0.37.2
sympy==1.12
tenacity==8.4.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
undetected-playwright==0.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.25.3
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.2
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.4
